Add EntityFixup class to VMFLib
This keeps track of the various bits of data associated with each fixup
value (the index, and original var name).
The methods on the class allow treating it as a dictionary, while it
ensures correct indexes are set and that keys can be checked
case-insensitively.
We then get rid of the get_fixup, set_fixup, has_fixup, etc methods and
just use regular indexing on the .fixup property.