Correct commands and streamline insructions
The command for compiling BEE2 itself fails to compile as it doesn't detect the compiler. Setting the `--distpath` value to `../dist/64bit/` instead of `/dist/BEE2/` corrects this.

Additionally, there is an unclear instruction to move the compiler to another folder. As the new command negates the need to move any files/folders around, it has been removed for simplicity.