Update FAQ.md
Removed secondary reason for not implementing toggleable light sources.
I don't think using lighting as a testing element is the primary purpose for it. Instead, the uses, which could be quite creative, are purely aesthetic. Most triggers would simply be entering an area. Others could be connected to test element inputs as special effects. I'm proposing its removal because it assumes a lack of good map making skills and isn't (in my perspective) a very good argument.